tpch: support int64 orderkey
With scale factors greater than 300, the l_orderkey column in the lineitem
table exceeds the capacity of a 32-bit integer. This patch increases it to
64-bit.

I also changed it to a signed integer, because we plan to eventually remove
unsigned ints. Thus, it makes more sense to benchmark the signed variant.

I also added better error messages to CHECK failures, since the overflow wasn't
obvious to diagnose.

Change-Id: Ib2ab10b1e9d244c0317792bcb8b003b946125709
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6438
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins