gutil: typedef 'uint64' to 'uint64_t' (and friends)
For whatever reason, the compiler was previously considering 'uint64' and
'uint64_t' to be different types due to the way they are defined -- probably
that 'uint64_t' is defined in standard headers to some internal __uint64 type,
whereas 'uint64' was defined by integral_types.h to be 'unsigned long long'.

This was annoying when trying to use certain gutil functions like SimpleItoa()
with a uint64_t (which we use most places in our code).

This patch changes integral_types.h to typedef the "google" versions of the types
to match the stdint ones.

Change-Id: Icccbd6cf11de526068f2246e4ffee8a46d1aed94
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6437
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins