KUDU-966. Add tracing for some network-related functions
My current theory is that KUDU-966 (slow calls to GetMasterRegistration)
is because this call ends up calling GetFQDN(), which calls getaddrinfo()
which might be slow in the case of a broken DNS setup.

This adds traces and LOG_SLOW_EXECUTION in the relevant code paths to help
diagnose such an issue. I also added traces to other non-perf-critical methods
in the same area of the socket/networking code.

Change-Id: Id4892eb9e7f256811504cb813e539107c765700d
Reviewed-on: http://gerrit.cloudera.org:8080/2051
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins