memtracker: unified hierarchy and bugfix extravaganza
I began with a simple problem: how to plumb a MemTracker into a server's
FsManager? This turns out to be hard because there's no server-wide
tracker (only per-tablet ones), so I set out to fix that. Such a tracker
should be constructed in ServerBase and exposed through various components
as needed.

As I worked through this, the requirement that every tracker have a unique
ID quickly grew painful (especially in minicluster tests), so I decided to
rewrite the static tracker handling to make it both more flexible and
intuitive. The idea is to store the tracker hierarchy just like it's
modeled: as a tree. There's a "root" tracker to which all other trackers
must be parented. This true tree hierarchy makes it easy to relax tracker
ID uniqueness: IDs must only be unique in the context of a given parent.
But it also means we can no longer pass tracker IDs down the stack as
parents; we have to pass the trackers themselves down.

With the tracker hierarchy reworked, plumbing a server-wide tracker became
simple. And as written, it gracefully handles the minicluster case by
producing trackers with names like "server 1", "server 2", etc.

Along the way, I fixed a number of issues I found, some of which were bugs
and some of which I plain didn't like:
- There can be multiple outstanding DMS trackers in each rowset, but each
  must be unique for the tablet. I added the rowset ID to the tracker name
  to fix this.
- Caches are singletons, so let's treat their trackers as such.
- The use of FindTracker() in the MemTracker constructor leads to recursive
  lock acquisiton when using FindOrCreateTracker() with a parent tracker.
- MemTrackers now have ToString(), which produces the entire tracker chain
  up to the root (and is thus globally unique).
- CreateTracker() will now properly report duplicate trackers (though again,
  the idea of a "duplicate" has been relaxed).
- All trackers are now auto-unregistered. Not sure why this wasn't default
  behavior from the get go.
- Trackers release their consumption when destructed. Arena-based trackers
  already did this in the arena destructor, but other trackers do not.

Change-Id: Iba7d8baa03a15c8fc9ca86c93747cbe93a203462
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6458
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>