external_mini_cluster: allow overriding webserver interface
This is handy when running something like tpch_real_world on a
remote system, and you want to be able to connect to the minicluster
from your desktop. Moving the position to before the "extra_flags"
allows you to override it (e.g. using the tpch_real_world flag file).

Change-Id: I248492f8673d85b3d726f3422e7d57c0e78bafcd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6460
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins