KUDU-299: [java client] Use a builder-like API to configure MR I/O formats
This patch changes the current static methods to adopt a builder-style API. Each branch of
configurators needed to be split in two because you can either pass the master quorum and op
timeout directly or you can rely on the CommandLineParser.

I did a few iterations on this design and it seemed the best one that directs the user into
using the API correctly without having to rely on asserts. OTOH, it seems over-engineered. Open
to suggestions.

Change-Id: I5b898038652c94026f45e61261b2319c934670c3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6463
Reviewed-by: Andrew Wang <andrew.wang@cloudera.com>
Tested-by: jenkins