KUDU-269. Add support for float/double
This adds FLOAT and DOUBLE types to the C++ client, the servers, a couple of
tests, and to the TPCH code.

Tested with a couple unit test additions, and the fact that the TPCH program
exercises them. I manually verified the results of TPCH Q1 vs the TPC-H spec.

I0423 18:27:01.043992 21694 tpch1.cc:193] Result:
I0423 18:27:01.044026 21694 tpch1.cc:204] A, F, 37734107, 56586554400.73, 25.52, 38273.13, 0.05, 1478493
I0423 18:27:01.044072 21694 tpch1.cc:204] R, F, 37719753, 56568041380.90, 25.51, 38250.85, 0.05, 1478870
I0423 18:27:01.044085 21694 tpch1.cc:204] N, O, 74476040, 111701729697.74, 25.50, 38249.12, 0.05, 2920374
I0423 18:27:01.044093 21694 tpch1.cc:204] N, F, 991417, 1487504710.38, 25.52, 38284.47, 0.05, 38854

In the process of implementation, I noticed that the code in row_util to
stringify a row was duplicated, so I removed it.

Handling float/double keys is not supported. Currently, attempting to use these
causes a crash trying to instantiate a KeyEncoder. I filed KUDU-725 to make
this behavior better.

Change-Id: I2d07bda1bfb69fe227c047043c98b87b46fe98fe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6472
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>