KUDU-617: track block manager memory usage
Now that we have a server-wide MemTracker, this commit plumbs it through the
FsManager and into the BlockManagers where a per-manager child tracker is
created and used to track "interesting" memory allocations. That boils down
to allocations that are either large or numerous (i.e. scale with amount of
data stored).

The file block manager only tracks ReadableBlock allocations, including
their embedded BlockIds and RandomAccessFiles.

The log block manager also tracks ReadableBlock allocations. Besides that,
it tracks LogBlocks (i.e. persisted block representations) as well as the
block map, which can become quite large as blocks are persisted.

Note that BlockId memory consumption is an approximation at best, because
the real value depends on whether the underlying std::string is reference
counted, and if so, whether we've already counted it or not. We could plumb
a MemTrackerAllocator all the way into the string, but it doesn't seem worth
the hassle when our approximation is reasonable.

Change-Id: I8b504079f2390a7ff040b53c7607e1b579e6f810
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6474
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>