[java client] Make TestLeaderFailover less flaky when looking for a leader
I've seen a test failure for TestLeaderFailover where we only found FOLLOWERs for a tablet when
looking for a LEADER. A few tens of ms later, the change was reported to the master. This patch
adds a retry loop with sleeping until we find a leader or timeout.

This patch also removes the hack where we used the CANDIDATE as a proxy to find the elected leader
since it's going away, and retry+sleep is a better solution anyways.

Change-Id: I57b0ac55868429f4b1831d175e4d8b4b35c7ec87
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6490
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>