Add a simple metrics logging thread to the server
This adds a thread which starts when the server is initialized that
periodically logs a metrics JSON dump.

For testing, I changed ExternalMiniCluster to enable this logging on all
servers with a 1-second interval. This is useful for a few reasons:
1) benchmarks that use miniclusters automatically get metrics logs, which
   we can use to generate graphs, etc.
2) all benchmarks that use miniclusters now cover metrics generation,
   which will help catch any bugs where FunctionGauge implementations
   lack thread-safety
3) it covers the new code

One result of point number 2 above is that I had to change
LogicalClock::NowForMetrics() so that it doesn't increment the clock.
Otherwise, the background polling of the metrics changed the logical clock
values expected by some tests.

A later commit in this series adds automatic gzipping of the log files
after each roll.

Change-Id: Ia0cf2774583dfa7eaa531629157b4dd99625438c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6494
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins