rolling_log: support compressing files after roll
This adds simple gzip compression of rolling log files after we roll to a new
one. Note that, rather than shelling out to 'gzip', I ended up implementing
a simple gzip function inline. The implementation comments detail the reasoning.

There is a simple test that verifies that the compressed log file shows up in
the log directory after a roll. The test doesn't attempt to decompress the file
-- it didn't seem worth it to implement gunzip for a unit test. Instead, it
just verifies that the gzip file is smaller than the original, but not empty.

I manually ran a tablet server with 42 tablets and the metrics logging interval
set to 100ms. Very quickly, I filled up the 64MB default rolling threshold, and
then that file got compressed down to 2MB:

-rw-rw-r-- 1 todd todd 2.0M Apr 28 23:42 /tmp/kudu-tablet_server.todd-ThinkPad-T540p.todd.metrics.20150428-234138.0.19909.gz
-rw-rw-r-- 1 todd todd 2.0M Apr 28 23:42 /tmp/kudu-tablet_server.todd-ThinkPad-T540p.todd.metrics.20150428-234217.0.19909.gz
-rw-rw-r-- 1 todd todd 2.0M Apr 28 23:43 /tmp/kudu-tablet_server.todd-ThinkPad-T540p.todd.metrics.20150428-234251.0.19909.gz
-rw-r--r-- 1 todd todd 8.2M Apr 28 23:43 /tmp/kudu-tablet_server.todd-ThinkPad-T540p.todd.metrics.20150428-234324.0.19909

Change-Id: Icea1aa2bec3167f5e8dca1189710675fbea0cec8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6495
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins