KUDU-628. Fix checksum calculation RPCs to do more work per RPC
This changes the server-side implementation of checksum scans to not return
early based on the amount of data scanned. The actual response size of a
checksum scan is constant size, so scanning more data per RPC doesn't risk
blowing out any buffers. Given that, it makes sense to spend a reasonable
amount of time on each RPC, so that the client doesn't get bound on
RPCs/second (a problem we see on the YCSB cluster).

With the fix, each checksum RPC responds after 500ms. I tested by using
tpch_real_world to insert 1GB of lineitem data in a local tablet server
and calling ksck -checksum_scan against it. During the checksum, I collected
tracing on the server side. Before the patch, this required 699 round trips to
the local server (in about 3.1 seconds). After the patch, it required 6 (and
took about 2.6 seconds). The resulting traces are available on the JIRA.

Change-Id: I31342187b4890bbc3725714dce8f569b342dfce8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6497
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins