KUDU-628. Further optimization of checksum scanning
This changes the checksum scan calculation to load each row into a faststring
before calling CRC on it. The CRC implementation is more efficient when running
on larger input strings, since it uses 64-bit-wide instructions (crc32q).
Additionally, the CRC implementation is in an external library so can't be
effectively inlined.

By putting the data to be checksummed in a faststring first, we amortize the
overhead of the non-inlined call, and we enable the wider checksum
instructions.

I benchmarked by running ksck on a SF=1 TPCH lineitem table locally. I ran
the checksum twice to ensure that the data was fully in block cache.

Before: 2.602s
After: 1.894s (27% improvement)

Change-Id: I626875e3f32ae52f8fb365d012665a9c892bd3b5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6498
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins