KUDU-735. Fix crash trying to read log entries towards the end of in-progress log
This fixes a crash we saw on the YCSB cluster in the following scenario:
- a log segment is preallocated to 64MB. Thus ReadableLogSegment::size_ is 67108864
- we reach the end of the log, but the async preallocation is a bit slow. This
  causes the last few records of the log to go past the preallocated range,
  yielding a log whose actual on-disk size is 67566660 bytes.
- we try to read one of those entries at the end of the log in the range that was
  not originally pre-allocated. This crashed with a "Truncation" error.

Previously, we were incorrectly comparing the range to be read against
file_size() which was an out-of-date measure of the size, which was only set
when the log segment was first opened. This happened to work OK because most
reads fell under the preallocation size. In some other scenarios, it didn't
work OK, but that was only due to a separate bug: we were using uint64_t for
offsets/sizes which meant that some of the comparisons ended up underflowing
and becoming ineffective.

This patch has a few fixes:
- when we append to a segment, we should also update the file_size_
- when checking whether a given read is valid, we should be using readable_up_to
  instead of file_size anyway.
- we should use int64_t for offsets and sizes in the code to make the underflow
  bug less likely (in fact just changing signedness caused the existing tests to
  detect this bug)
- fix the crash message we saw so that it would have been more useful

Change-Id: Ib5d61ddfa3a9980168bc25b7e800a79e982e4c91
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6555
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>