faststring: add ASAN poisoning
This adds manual poisoning to the faststring buffer for bytes which have been
allocated but are not part of the "valid" part of the data.

This caused a couple of test failures, but nothing looks like a real bug. We
were just being lazy and writing into "reserved" data in some places.

Nevertheless, it's probably a good safe-guard in case we are ever accidentally
overrunning a buffer but getting "lucky" and hitting reserved memory.

A seemingly-unrelated part of this patch is to add -DLLVM_SUPPORT_VALGRIND_H to
the compilation of the codegen module. The reasoning is that faststring.h now
includes dynamic_annotations.h, which conflicts with llvm/Support/Valgrind.h.
Adding this define prevents the inclusion of LLVM's version.

Change-Id: Icc8753745784251dd76e6e4bdbc4b4734d84eb25
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6557
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>