codegen: precompiled.ll should not be optimized
The precompiled LLVM IR should not be optimized during the build. Impala found
that, in some cases, optimizing the IR at compile time reduced the opportunity
for later optimization at runtime.

I looked at the resulting runtime assembly generated by
full_stack-insert-scan-test and it actually did not change as a result of this
fix. But, given their experience, I figured it was worth doing.

Change-Id: I9cd11cd997b6a3751c0ef06e3e4bab266ee2b934
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6563
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>