KUDU-435: Add a synchronous version of AsyncKuduClient
We currently only have one real synchronous interface and it's KuduSession, AsyncKuduClient does
contain some synchronous methods but that's it.

This patch creates a new KuduClient that's fully synchronous and that wraps an AsyncKuduClient. It
replicates all the methods from the latter, except for scanner-based methods that will be changed
in a follow-up patch.

The most impacted class was ITBLL since it uses the AsyncKuduClient in both synch and async ways.
This patch only tries to modifies the clients when necessary without having to use KuduClient
_and_ AsyncKuduClient within the same job.

Change-Id: I8550e235a3527b6ebb2be55595afe71ff60211ca
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6561
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>