codegen: fix LLVM parse errors in master process
We were seeing that some processes (such as the master) were emitting a warning
that code could not be generated because of a parse error. After much gdb'ing,
I discovered that LLVM actually expects the input buffer to be passed in with a
length that does not include the null terminator, but requires that the byte
_after_ the buffer is null.

This doesn't add any new unit test, since the failure before was dependent on
link order and a lot of other things that are tough to control. However, I
verified that the master no longer emits the error.

This patch also slightly improves the LLVM error message when it fails to
parse.

Change-Id: Ib13ad2dc35a7aa2a2ac5aeb2a87dd274a1512c8f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6562
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>