KUDU-712 (part 1): metrics prototypes should specify which entity they belong to
This adds metadata to the metrics prototypes which specifies the metric entity
type that they must be attached to. At runtime, we check that the entity
matches the expected type.

This is necessary for exporting an MDL for CM to consume: we need to know which
kinds of metrics to expect attached to which kinds of entities.

It's worth noting that the mapping is achieved using strings rather than direct
pointers to MetricEntityPrototype objects. I originally tried that other approach,
but ran into link issues: for example, consensus/log_metrics.cc needs to define
some metrics which are attached to 'tablet', but the tablet entity is defined in the
'tablet' module. Adding a dependency between these two modules would likely create
a cycle. Using a string is sufficient as all necessary validity checks already happen
at runtime.

No new unit test is added, because all of the metrics code paths exercise this
code already. A couple unit tests had to be modified so the local metric entities
they attached objects to matched the expected entity types.

Change-Id: Ic9a1ea37a675a41ebf4fd16f8751ea0928c497b8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6565
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>