KUDU-435: Add a synchronous scanner
The last class that needed a sync version was KuduScanner, which is now renamed AsyncKuduScanner
and a new KuduScanner with a sync API replaces it.

The one tricky thing in this patch is the scanner builder which required a different ? super T
incantation. It's very much like the KuduTableMapReduceUtil configurators, but not quite.

This changes also a few more tools to use the new KuduScanner.

Change-Id: Iaa73e37d9f80090ca7a0d771bc6f147e606642c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6585
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>