KUDU-712 (part 2): add --dump_metrics_json to dump all of the metrics prototypes
This adds a flag which is implemented across all Kudu processes that dumps the
metrics prototypes in JSON format. The format is not MDL-compatible. I plan to
do some post-processing in a python script to massage the JSON into the format
expected by CM.

In addition to the new unit test, I manually verified that running
'kudu-master --dump_metrics_json' dumped a reasonable looking JSON document.

Change-Id: I430f6fb012e5930360febd1569e7c7cc9d563ae5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6579
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins