fs: parallelize opening of the log block manager
On the 4.4 TB deployment, opening the log block manager took 10-15s. And
that's a cost we can't amortize across multiple tablets. We can do better if
we parallelize the work across each disk.

Note that there's still a great deal of contention on the block manager
lock; that will be addressed in a follow on patch.

Change-Id: I957eebd3171646ec78f8c6765c3e52c9b69aa558
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6587
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>