Add a script to convert metrics log to TSV
I've been using this script in conjunction with R to plot graphs of various
metrics from tpch_real_world runs. It's not yet meant to be a completed
user-facing tool (eg the set of metrics to extract are hard-coded) but it's
useful to check in so that others can use it as well.

Change-Id: Ib495f1b32d5717ec13e3ff27fe2a9fb759a2bd46
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6591
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>