KUDU-712 (part 3): add attributes to metric entities
This adds an attributes map to the MetricEntity class, and provides some simple
APIs to set them, either when the entity is created or at some later point in
time.

This will be used by CM to do hierarchical aggregation of metrics (e.g. rolling up
all tablet metrics on a per-table basis).

Change-Id: I8e5b5784cb4c193d4cac88a7158f91f92a4176d1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6595
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>