Move test helpers into cluster_itest_util.h
We currently rely too much on these test-base classes that are hard to
reuse in different contexts. There are some functions I want access to
from a new test that is using MiniCluster, while the helpers assume
ExternalMiniCluster for no good reason. There are also some other
features missing from MiniCluster so pull those in as well.

Changes:
 - Create a cluster_itest_util.h (and .cc file) for integration tests to
   share helper functions and classes in.
 - Factor out TServerDetails class into cluster_itest_util.h from
   ts_itest-base.h
 - Also factor some helper functions out of ts_itest-base.h:
   - GetLastOpIdForEachReplica()
   - WaitUntilAllReplicasHaveOp()
   - WaitForServersToAgree()
   - CreateTabletServerMap()
 - Factor CreateTsClientProxies() out of tserver/tablet_server-test-base.h
   into tserver/tablet_server_test_util.h so that it can be called by
   CreateTabletServerMap()
 - Add CreateClient() to MiniCluster

In order to extricate and reuse TServerDetails in a MiniCluster context,
I had to remove the pointer to ExternalTabletServer from the struct. To
make up for this, I added an accessor to ExternalMiniCluster to get an
ExternalTabletServer by UUID, which is stored in the TServerDetails
NodeInstancePb object. Then I tweaked the two tests that relied on that
functionality to work with the new setup.

Change-Id: Ib9496507ba04097eed30fb2a7472ccea35d481f0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6598
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>