KUDU-746: Add datadirs, safety valves, and icon to Kudu CSD
This commit will add the ability to configure datadirs in the CM wizard, as
well as safety valves for the master and tablet servers.

The safety valves will be used to add additional command-line options, in
gflags format (e.g. "--master_num_acceptors_per_address=2).

I forgot to add the icon from KUDU-696, so will add here as well.

Change-Id: I5cba5b869bc3c079ac69fa52f1d55d75fdc16cf1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6597
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Vikram Srivastava <vikrams@cloudera.com>