KUDU-749: improve performance of applying deltas from DMS
After running YCSB in a zipfian 50/50 R/W workload, I saw a lot of CPU
usage in the delta application path from DMS. The culprits of the CPU
usage are manyfold:

The current implementation buffers all of the deltas in a faststring. We were
using faststring::reserve() to reserve exactly the right amount for the each
delta that we saw. This prevented the faststring from doing its normal doubling
behavior. Rather, every delta that we appended caused a memcpy() of all
previous deltas, resulting in O(n^2) runtime.

I added a new benchmark which applies 10,000 updates to a single row and times
how long it takes for them to be applied to a column block. With the original
code, this is incredibly slow:

  I0509 14:27:31.330876 16668 deltamemstore-test.cc:203] Time spent Applying updates: real 5.812s user 5.772s     sys 0.040s
  * Modified the benchmark to not loop 100 times. So, the true number is 581sec
  with the test loop enabled.

After fixing the faststring::reserve() call, it improved substantially:
  I0509 14:28:26.919739 16839 deltamemstore-test.cc:203] Time spent Applying updates: real 0.705s user 0.688s     sys 0.016s
(savings of 830x!)

However, when I tested this on the cluster, the DMS application was still a CPU hog.
There are two more issues that are more fundamental:

- We pass through all of the deltas once for each column. This makes delta application
  O(num_deltas * num_columns) instead of the preferred O(num_deltas). This is true
  even if only a fraction of the columns are getting updated!

- Even if a single column is updated multiple times, we copy each of the intermediary
  values into the destination arena. This is very costly when the values are large
  (eg the 100byte binary strings in YCSB). Instead, we should try to copy only the
  very latest value into the destination arena.

The solution taken by this patch is to change the PrepareBatch() implementation
to parse the deltas and separate them into arrays for each column. While doing so,
it becomes trivial to allow later updates to overwrite earlier ones without copying
the referenced data into the target arena.

Once the updates are separated by column, we only need to pass through the target
column when actually applying updates. This results in the intended O(n_updates)
behavior at the time of ApplyUpdates.

 Performance counter stats for './build/latest/deltamemstore-test --gtest_filter=*Bench* --benchmark_num_passes=10000' (5 runs):

       6030.851298 task-clock                #    0.991 CPUs utilized            ( +-  0.38% )
               489 context-switches          #    0.081 K/sec                    ( +- 13.50% )
                19 cpu-migrations            #    0.003 K/sec                    ( +- 22.78% )
             6,723 page-faults               #    0.001 M/sec                    ( +-  0.02% )
    19,922,357,444 cycles                    #    3.303 GHz                      ( +-  0.40% )
   <not supported> stalled-cycles-frontend
   <not supported> stalled-cycles-backend
    43,254,862,336 instructions              #    2.17  insns per cycle          ( +-  0.09% )
     8,673,378,065 branches                  # 1438.168 M/sec                    ( +-  0.00% )
        27,818,121 branch-misses             #    0.32% of all branches          ( +-  0.08% )

       6.087171528 seconds time elapsed                                          ( +-  0.44% )

Note that the time here is for 10,000 passes, whereas the original time (5.8s) quoted up above
is for 1 pass. So, the total improvement here is approximately 10,000x.

Change-Id: Ib00b8ba613ef58094bcd0a61a66ffdc93e85b6cb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6603
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>