java: fix bugs in string handling in operation
This fixes a number of bugs related to STRING column type handling:
- we were conflating the String.length() with the length of the UTF8-encoded
  string. This would prevent us from storing non-ASCII strings.
- if the strings were presented out of order, we incorrectly assigned
  strings to columns in the order that they were set, not to the correct
  columns.
- if we set a string and then set it back to NULL, we'd end up with a mis-sized
  buffer.

Even aside from strings, the buffer size management did not handle the case where
a column was set multiple times before applying.

In order to fix this, I ended up getting rid of the code which tries to estimate
the buffer size requirements as the row is being built. It was too complex to get
right, as evidenced by the above bugs. Instead, I just use an automatically growing
buffer as the response is being serialized. This has the downside of adding at least
one extra buffer copy, but I figured it was better to restore correctness and
re-optimize later, rather than try to both fix the bugs and retain speed.

This adds a few unit tests for insert encoding as well as a functional test which
inserts and retrieves string data.

While I was at it, I added a new public API which allows the user to directly
insert an already-UTF8-encoded string. I noticed that YCSB spends a majority
of its CPU doing UTF8-encoding, but the actual input strings are just ASCII.
So, we can take some load off the YCSB client by using this new API. I imagine
that other consumers of the API can make similar optimizations in many cases.

Change-Id: Idb470e0e90f9b3cf9a95903fd1e8439e85a6a34e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6625
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>