Fix prioritization of DMS flushes
Previously, if two DMSs had the same amount of retained logs, we'd always
prefer to flush the one we visited first, regardless of how much in-memory
data we had. In a zipfian update workload, I saw that this was causing some
DMS towards the end of the list to get flushed only very rarely.

Instead, if two DMS have the same retention, we should prefer to flush the
one that has more data.

I tested this on bolt80 under several YCSB random update workloads (both
zipfian and uniform). It seemed to help, but the maintenance manager is
a little bit opaque right now (further patches to follow to make it a little
more debuggable)

Change-Id: I2db5ee6e9a882bb40832dce3fa0dd910d41c1d66
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6607
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>