Add VLOG_AND_TRACE(...) macro
This adds a macro which writes to VLOG(n) and also to the
current trace event buffer.

My initial use case is that I'd like to see a bit of logging about
the maintenance manager decision-making process within a trace
output. We already have VLOG(1) for these, so this macro will make
it easy to also expose the info in the trace.

Change-Id: I3f75c2997b0e0061664fa52433e9ce7eb72bcd11
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6606
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>