ts-cli: add ability to dump a tablet
This adds kudu-ts-cli dump_tablet, which dumps the rows of a tablet replica on
a specific server.

I built this in order to diagnose some replica divergence issues I was seeing
on bolt80. I used it there and it worked great to dump the rows of the replicas
in order, allowing me to use 'diff' later to see how they had diverged.

Change-Id: Idb974c0150991254f871d2639fbeaa30dd39ce62
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6615
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins