java: make tests automatically discover binaries and build paths
This allows the Java tests to be run from within Eclipse without
any special setup:

Automatically find the build directory by traversing upwards to find the
'.git' in the root of the repository. This can be overridden with a system
property.

Find the 'flags' file by making it a Java "resource" in src/test/resources
instead of using the non-standard "test-resources" directory. In the case
that we're running within a jar, we extract the flags file into a tmp location.

Uses reasonable defaults (localhost cluster) when properties are not found.

Change-Id: Ia1c56a52016ed3ddcfe63fe7bf03dbbc841ce1ec
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6624
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>