KUDU-765. New master leaders should wait for old ops to commit
When there is a master failover, the new master will visit its tablets
as soon as it gets elected leader. This can be a problem because it may
not have received the COMMIT message for a committed Raft message, which
means it can essentially "travel back in time" for a while, and even
worse, it may never re-visit the sys catalog so will "forget" things
that are on disk because it assumes it had caught up to the
previously-committed state and basically leave a hole in its knowledge.

To fix this, we wait until at least one op from the current term is
committed.

This was the CHECK failure that alerted me to this issue. It came from
the TestMasterFailover#testKillLeader unit test in the java client.
After applying the patch to switch CC ops for NO_OP ops when asserting
leadership.

F0512 19:46:03.614502 22251 catalog_manager.cc:1307] Check failed: tablet_lock.data().pb.state() == SysTabletsEntryPB::kTabletStateCreating
*** Check failure stack trace: ***
     @     0x7f8357161bad  google::LogMessage::Fail()
     @     0x7f8357166047  google::LogMessage::SendToLog()
     @     0x7f8357163ea9  google::LogMessage::Flush()
     @     0x7f83571641ad  google::LogMessageFatal::~LogMessageFatal()
     @     0x7f835746ef63  kudu::master::CatalogManager::HandleReportedTablet()
     @     0x7f835746d651  kudu::master::CatalogManager::ProcessTabletReport()
     @     0x7f83574b1bb6  kudu::master::MasterServiceImpl::TSHeartbeat()
     @     0x7f8351f56f72  kudu::master::MasterServiceIf::Handle()
     @     0x7f8351c7a6ba  kudu::rpc::ServicePool::RunThread()
     @     0x7f8351c7c9c5  boost::_mfi::mf0<>::operator()()
     @     0x7f8351c7c784  boost::_bi::list1<>::operator()<>()
     @     0x7f8351c7c627  boost::_bi::bind_t<>::operator()()
     @     0x7f8351c7c4ad  boost::detail::function::void_function_obj_invoker0<>::invoke()
     @     0x7f8351c3412b  boost::function0<>::operator()()
     @     0x7f8350f6cd3a  kudu::Thread::SuperviseThread()

Additionally, this patch does a little bit of cleanup by avoiding
holding a reference to a TabletInfo object in the AsyncCreateTablet
class.

Also reduce the timeout for the testKillLeader() test from 100 seconds
to 30 seconds (which is still much longer than it should take).

Change-Id: I981ee43fe52740cc12d9a5c2f2fd0bd1e3d2ad27
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6627
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins