KUDU-743. Fix case where client can mis-order operations within a batch
ClientTest.TestSeveralRowMutatesPerBatch was flaky because there was a race
where the client could re-order operations within a batch before sending them
to the server.

I didn't carefully analyze the race, but it seems likely that it can happen if
the first operation causes a location lookup, and just as that lookup
completes, the second operation is added to the batch. Then, the second
operation might be put into the per-TS buffer ahead of the first one.

The solution here is relatively simple: as operations are added to the batch,
they are assigned a sequence number. Then, when they are added to the per-TS
buffer, we do a sort of "reverse bubble sort" to ensure that any re-orderings
are resolved.

Without the patch, looping the above-mentioned test ~100 times usually causes
a failure. With the patch, I looped several hundred times and did not see
a failure.

Change-Id: Ic5982069db45da2be5ea0ef5cad405df0460d7cb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6645
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>