KUDU-445: First cut at a Python API client.
Start building Cython-based wrapper interface to the Kudu C++ client
API. This wrapper is treated as highly experimental. Currently the
folloing simple features work:

- Create schemas and create / delete tables
- Tested simple inserts and delete
- Implemented Scanner and tested basic range predicates
- Can convert a set of KuduRowResult objects to Python tuples
- Translate client KuduError to Python-readable form
- Translate internal Kudu Status messages to Python exceptions

This patch adds as well an integration to Jenkins to make sure that the
Python API is not falling behind the C++ API.

Original prototype authored by Wes McKinney.

Change-Id: I3818dee1de886ba362d504d0664600b51a849793
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5196
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>