Encoded keys in TabletStatusPB should be 'bytes', not 'string'
I noticed that using kudu-ts-cli against a server in bolt80 that it
spewed the following warnings:

[libprotobuf ERROR google/protobuf/wire_format.cc:1053] String field contains
invalid UTF-8 data when parsing a protocol buffer. Use the 'bytes' type if you
intend to send raw bytes.

Indeed it's right - we should use 'bytes' for encoded keys.

There's no test for this since the protobuf error is just a warning on stderr,
nothing that can be asserted on.

Change-Id: I101b41304701e3dc712e8cc5e520a0d7808061ab
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6664
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>