Initial cut of distributed testing tools
These tools allow a developer to submit tests to be executed on a
remote testing service.

This depends on the 'dist_tests' repository[1] and currently only
work if the testing service is running a compatible version of libc
and libstdc++ (eg submitting from el6 to el6, not from ubuntu to el6)

Further documentation to follow in a later commit and/or on the wiki.

[1] git@github.mtv.cloudera.com:todd/dist_test.git

Change-Id: I47d98e1824e3e8979400905b9d538eaf058eb8a0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6692
Reviewed-by: Andrew Wang <andrew.wang@cloudera.com>
Tested-by: jenkins